+++ src/Controller/Admin/PostsController.php
@@ -55,13 +55,13 @@ class PostsController extends AppController
             $users = $this->Posts->Users->getList();
         }
 
-        if (!$users) {
+        if ($users->isEmpty()) {
             $this->Flash->alert(__d('me_cms', 'You must first create an user'));
 
             return $this->redirect(['controller' => 'Users', 'action' => 'index']);
         }
 
-        if (!$categories) {
+        if ($categories->isEmpty()) {
             $this->Flash->alert(__d('me_cms', 'You must first create a category'));
 
             return $this->redirect(['controller' => 'PostsCategories', 'action' => 'index']);